refactor(tests): split tests to module transactions and predicates
