+++ .env
@@ -1,0 +1,6 @@
+ENCODER='fuel',
+API_URL='http://localhost:3333',
+PROVIDER='http://localhost:4000/graphql',
+GAS_PRICE=1_000_000,
+GAS_LIMIT=100000,
+REFETCH_TIMEOUT=1000
