+++ tsconfig.json
@@ -16,5 +16,5 @@
         "target": "es6"
     },
     "include": ["src/**/*.ts", "src/**/*.tsx"],
-    "exclude": ["node_modules", "src/__tests__/*.ts", "src/mocks/*.json"]
+    "exclude": ["node_modules", "src/library/predicates/__tests__/*.ts", "src/mocks/*.json"]
 }
